requirements.txt
flask==2.3.2
pandas==2.0.3
sklearn=1.3.0
flask-mail==0.9.1
python-dotenv==1.0.0
mysql-connector-python==8.0.33

